Refactor login endpoint to use strategy
