+++ src/server.ts
@@ -1,6 +1,7 @@
-import express from "express";
+import express, { Request, Response } from "express";
 import dotenv from "dotenv";
 import rateLimit from "express-rate-limit";
+import { Passport } from "passport";
 
 import { getNearbyPlaces } from "./rest/getNearbyPlaces";
 import { getPlacePhoto } from "./rest/getPlacePhoto";
@@ -14,7 +15,8 @@ import { addPlaceToDb } from "./rest/Places/addPlace";
 import { editRating } from "./rest/Ratings/editRating";
 import { editUser } from "./rest/Users/editUser";
 import { deleteRating } from "./rest/Ratings/deleteRating";
-import { login } from "./rest/login";
+import { strategy } from "./rest/util";
+import StatusCode from "./rest/status";
 
 export const app = express();
 const port = Number(process.env.PORT) || 3030;
@@ -25,9 +27,13 @@ const limiter = {
 		"The API is rate limited to a maximum of 1000 requests per 15 minutes, please lower your request rate",
 };
 
+const auth = new Passport();
+auth.use(strategy);
+
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(rateLimit(limiter));
+app.use(auth.initialize());
 
 dotenv.config();
 
@@ -70,7 +76,14 @@ app.patch("/editUser", editUser);
 // delete rating
 app.delete("/deleteRating/:id", deleteRating);
 
-app.post("/login", login);
+app.post(
+	"/login",
+	auth.authenticate(strategy, { session: false }),
+	(req: Request, res: Response): void => {
+		const token = req.user;
+		res.status(StatusCode.OK).send(token);
+	}
+);
 
 export const server = app.listen(port, () => {
 	console.log(`App listening at http://localhost:${port}`);
